move stuff and restructure isomorphism checker
